requirement.txt
pandas
plotly_express
streamlit
numpy
